INT-1618: Refactor Ferb API to return queries in batches - refactor ScreeningDao (move method related to IntakeLOVCodeEntity into a separate DAO)
